[refactor] more organization of components
