+++ frontend/src/pages/home-page/Home.tsx
@@ -12,13 +12,13 @@ import {
 import "../../styles/globals.css";
 import Hero from "./components/Hero";
 import DashboardPreview from "../../components/Chart";
-import { Safari } from "../../components/ui/safari";
+import { Safari } from "./components/safari";
 import { AnimatedListHome } from "../../components/ui/animated-list/animated-list-home"; 
 import { Faq } from "../../components/ui/faq/faq";
 // import RiskTable from "../components/RiskTable";
 
 import HeroCard  from "./components/HeroCardUI";
-import { BentoCard, BentoGrid } from "../../components/ui/bento-grid";
+import { BentoCard, BentoGrid } from "./components/bento-grid";
 
 const features = [
   {
