+++ frontend/src/pages/home-page/components/bento-grid.tsx
@@ -1,7 +1,7 @@
 import type { ComponentPropsWithoutRef, ReactNode } from "react"
 import { ArrowRightIcon } from "@radix-ui/react-icons"
-import { cn } from "../../../lib/utils"
-import { Button } from "./button"
+import { cn } from "../../../../lib/utils"
+import { Button } from "../../../components/ui/button"
 
 interface BentoGridProps extends ComponentPropsWithoutRef<"div"> {
   children: ReactNode
