+++ frontend/src/pages/home-page/components/safari.tsx
